R: Move foo2 to another clas
