+++ StockTracker/StockTracker.csproj
@@ -46,6 +46,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Foo2Foo2.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
